livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[silero,turn-detector,openai]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "langchain-valyu",
    "langchain",
    "openai>=1.0.0",
]
